pyproject.toml
[project]
name = "reddit-avatar"
version = "0.1.0"
description = "Config-driven Reddit → avatar markdown report pipeline"
readme = "README.md"
requires-python = ">=3.11"
authors = [{ name = "tatlaoui", email = "tarik.atlaoui@square-sense.com" }]
dependencies = [
    "anthropic>=0.40",
    "httpx>=0.27",
    "jinja2>=3.1",
    "pydantic>=2.7",
    "python-slugify>=8.0",
    "pyyaml>=6.0",
    "rich>=13.7",
    "tenacity>=9.0",
    "typer>=0.12",
]

[project.scripts]
reddit-avatar = "reddit_avatar.cli:app"

[dependency-groups]
dev = [
    "pytest>=8.0",
    "ruff>=0.6",
]

[build-system]
requires = ["uv_build>=0.9.17,<0.10.0"]
build-backend = "uv_build"

[tool.ruff]
line-length = 100
target-version = "py311"

[tool.pytest.ini_options]
testpaths = ["tests"]
